+++ src/rules/no-extraneous-dependencies.js
@@ -126,6 +126,36 @@ function getModuleRealName(resolved) {
   return getFilePackageName(resolved);
 }
 
+function checkDependencyDeclaration(deps, packageName) {
+  // in case of sub package.json inside a module
+  // check the dependencies on all hierarchy
+  const packageHierarchy = [];
+  const packageNameParts = packageName.split('/');
+  packageNameParts.forEach((namePart, index) => {
+    if (!namePart.startsWith('@')) {
+      const ancestor = packageNameParts.slice(0, index + 1).join('/');
+      packageHierarchy.push(ancestor);
+    }
+  });
+
+  return packageHierarchy.reduce((result, ancestorName) => {
+    return {
+      isInDeps: result.isInDeps || deps.dependencies[ancestorName] !== undefined,
+      isInDevDeps: result.isInDevDeps || deps.devDependencies[ancestorName] !== undefined,
+      isInOptDeps: result.isInOptDeps || deps.optionalDependencies[ancestorName] !== undefined,
+      isInPeerDeps: result.isInPeerDeps || deps.peerDependencies[ancestorName] !== undefined,
+      isInBundledDeps:
+        result.isInBundledDeps || deps.bundledDependencies.indexOf(ancestorName) !== -1,
+    };
+  }, {
+    isInDeps: false,
+    isInDevDeps: false,
+    isInOptDeps: false,
+    isInPeerDeps: false,
+    isInBundledDeps: false,
+  });
+}
+
 function reportIfMissing(context, deps, depsOptions, node, name) {
   // Do not report when importing types
   if (node.importKind === 'type' || (node.parent && node.parent.importKind === 'type') || node.importKind === 'typeof') {
@@ -139,37 +169,49 @@ function reportIfMissing(context, deps, depsOptions, node, name) {
   const resolved = resolve(name, context);
   if (!resolved) { return; }
 
-  // get the real name from the resolved package.json
-  // if not aliased imports (alias/react for example) will not be correctly interpreted
-  // fallback on original name in case no package.json found
-  const packageName = getModuleRealName(resolved) || getModuleOriginalName(name);
+  const importPackageName = getModuleOriginalName(name);
+  const importPackageNameDeclaration = checkDependencyDeclaration(deps, importPackageName);
 
-  const isInDeps = deps.dependencies[packageName] !== undefined;
-  const isInDevDeps = deps.devDependencies[packageName] !== undefined;
-  const isInOptDeps = deps.optionalDependencies[packageName] !== undefined;
-  const isInPeerDeps = deps.peerDependencies[packageName] !== undefined;
-  const isInBundledDeps = deps.bundledDependencies.indexOf(packageName) !== -1;
-
-  if (isInDeps ||
-    (depsOptions.allowDevDeps && isInDevDeps) ||
-    (depsOptions.allowPeerDeps && isInPeerDeps) ||
-    (depsOptions.allowOptDeps && isInOptDeps) ||
-    (depsOptions.allowBundledDeps && isInBundledDeps)
+  if (importPackageNameDeclaration.isInDeps ||
+    (depsOptions.allowDevDeps && importPackageNameDeclaration.isInDevDeps) ||
+    (depsOptions.allowPeerDeps && importPackageNameDeclaration.isInPeerDeps) ||
+    (depsOptions.allowOptDeps && importPackageNameDeclaration.isInOptDeps) ||
+    (depsOptions.allowBundledDeps && importPackageNameDeclaration.isInBundledDeps)
   ) {
     return;
   }
 
-  if (isInDevDeps && !depsOptions.allowDevDeps) {
-    context.report(node, devDepErrorMessage(packageName));
+  // test the real name from the resolved package.json
+  // if not aliased imports (alias/react for example), importPackageName can be  misinterpreted
+  const realPackageName = getModuleRealName(resolved);
+  const realPackageNameDeclaration = checkDependencyDeclaration(deps, realPackageName);
+
+  if (realPackageNameDeclaration.isInDeps ||
+    (depsOptions.allowDevDeps && realPackageNameDeclaration.isInDevDeps) ||
+    (depsOptions.allowPeerDeps && realPackageNameDeclaration.isInPeerDeps) ||
+    (depsOptions.allowOptDeps && realPackageNameDeclaration.isInOptDeps) ||
+    (depsOptions.allowBundledDeps && realPackageNameDeclaration.isInBundledDeps)
+  ) {
     return;
   }
 
-  if (isInOptDeps && !depsOptions.allowOptDeps) {
-    context.report(node, optDepErrorMessage(packageName));
+  if ((
+    importPackageNameDeclaration.isInDevDeps ||
+    realPackageNameDeclaration.isInDevDeps
+  ) && !depsOptions.allowDevDeps) {
+    context.report(node, devDepErrorMessage(realPackageName));
     return;
   }
 
-  context.report(node, missingErrorMessage(packageName));
+  if ((
+    importPackageNameDeclaration.isInOptDeps ||
+    realPackageNameDeclaration.isInOptDeps
+  ) && !depsOptions.allowOptDeps) {
+    context.report(node, optDepErrorMessage(realPackageName));
+    return;
+  }
+
+  context.report(node, missingErrorMessage(realPackageName));
 }
 
 function testConfig(config, filename) {